+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CDP Python SDK Changelog
 
 ## Unreleased
+- Add support for fetching address reputation
+  - Add `reputation` method to `Address` to fetch the reputation of the address.
 
 ### Added